+++ lib/src/compatibility/providers/openai_family_compat_openai_config.dart
@@ -4,7 +4,7 @@ import '../config/legacy_config_extensions.dart';
 import '../config/legacy_openai_options.dart';
 import '../config/legacy_provider_options.dart';
 import '../config/legacy_web_search_options.dart';
-import 'community_provider_config_adapters.dart';
+import 'legacy_dio_client_overrides.dart';
 
 OpenAIConfig createLegacyOpenAIConfig(LLMConfig config) =>
     toCompatLegacyOpenAIConfig(config);